project_root/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.8
django-environ==0.12.0
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
gunicorn==23.0.0
kombu==5.6.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
